🚚 Move is_valid_* to new crate zlink-names
This commit moves:
- is_valid_field_name
- is_valid_interface_name
- is_valid_type_name

This also moves/renames *_name functions to parse_*_name.
