+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "zlink-codegen",
     "zlink-codegen/test-integration",
     "zlink-idl",
+    "zlink-names",
     "zlink-tokio",
     "zlink-smol",
     "zlink-macros",
