+++ zlink-names/Cargo.toml
@@ -1,22 +1,20 @@
 [package]
-name = "zlink-idl"
+name = "zlink-names"
 version = "0.7.0"
-description = "Varlink IDL types and parser"
+description = "Varlink name types"
 edition.workspace = true
 rust-version.workspace = true
 license.workspace = true
 repository.workspace = true
 authors.workspace = true
 
-[features]
-parse = ["dep:winnow"]
-
 [dependencies]
+regex = { version = "1.13.1", default-features = false }
 winnow = { version = "1.0.4", default-features = false, features = [
     "alloc",
-    "ascii",
-], optional = true }
-zlink-names = { path = "../zlink-names", version = "=0.7.0", default-features = false }
+    "parser",
+] }
+zcheapstr = "1.1.0"
 
 [package.metadata.docs.rs]
 all-features = true
